+++ src/Symfony/Component/Messenger/Tests/TraceableMessageBusTest.php
@@ -27,7 +27,7 @@ class TraceableMessageBusTest extends TestCase
 
         $stamp = new DelayStamp(5);
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
-        $bus->expects($this->once())->method('dispatch')->with($message, [$stamp])->willReturn(new Envelope($message));
+        $bus->expects($this->once())->method('dispatch')->with($message, [$stamp])->willReturn(new Envelope($message, [$stamp]));
 
         $traceableBus = new TraceableMessageBus($bus);
         $line = __LINE__ + 1;
@@ -38,6 +38,7 @@ class TraceableMessageBusTest extends TestCase
         $this->assertEquals([
             'message' => $message,
             'stamps' => [$stamp],
+            'stamps_after_dispatch' => [$stamp],
             'caller' => [
                 'name' => 'TraceableMessageBusTest.php',
                 'file' => __FILE__,
@@ -63,6 +64,33 @@ class TraceableMessageBusTest extends TestCase
         $this->assertEquals([
             'message' => $message,
             'stamps' => [$stamp],
+            'stamps_after_dispatch' => [$stamp],
+            'caller' => [
+                'name' => 'TraceableMessageBusTest.php',
+                'file' => __FILE__,
+                'line' => $line,
+            ],
+        ], $actualTracedMessage);
+    }
+
+    public function testItCollectsStampsAddedDuringDispatch()
+    {
+        $message = new DummyMessage('Hello');
+        $envelope = (new Envelope($message))->with($stamp = new AnEnvelopeStamp());
+
+        $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
+        $bus->expects($this->once())->method('dispatch')->with($envelope)->willReturn($envelope->with($anotherStamp = new AnEnvelopeStamp()));
+
+        $traceableBus = new TraceableMessageBus($bus);
+        $line = __LINE__ + 1;
+        $traceableBus->dispatch($envelope);
+        $this->assertCount(1, $tracedMessages = $traceableBus->getDispatchedMessages());
+        $actualTracedMessage = $tracedMessages[0];
+        unset($actualTracedMessage['callTime']); // don't check, too variable
+        $this->assertEquals([
+            'message' => $message,
+            'stamps' => [$stamp],
+            'stamps_after_dispatch' => [$stamp, $anotherStamp],
             'caller' => [
                 'name' => 'TraceableMessageBusTest.php',
                 'file' => __FILE__,
@@ -94,6 +122,7 @@ class TraceableMessageBusTest extends TestCase
             'message' => $message,
             'exception' => $exception,
             'stamps' => [],
+            'stamps_after_dispatch' => [],
             'caller' => [
                 'name' => 'TraceableMessageBusTest.php',
                 'file' => __FILE__,